Move OfaRuntimeWrapper to types module
