+++ crates/relayer-cosmos/src/base/impls/relay.rs
@@ -14,7 +14,7 @@ use ibc_relayer_framework::base::one_for_all::traits::chain::OfaChainTypes;
 use ibc_relayer_framework::base::one_for_all::traits::relay::{OfaBaseRelay, OfaRelayTypes};
 use ibc_relayer_framework::base::one_for_all::types::chain::OfaChainWrapper;
 
-use ibc_relayer_framework::base::one_for_all::traits::runtime::OfaRuntimeWrapper;
+use ibc_relayer_framework::base::one_for_all::types::runtime::OfaRuntimeWrapper;
 
 use crate::base::error::Error;
