+++ crates/relayer-cosmos/src/base/types/chain.rs
@@ -1,5 +1,5 @@
 use alloc::sync::Arc;
-use ibc_relayer_framework::base::one_for_all::traits::runtime::OfaRuntimeWrapper;
+use ibc_relayer_framework::base::one_for_all::types::runtime::OfaRuntimeWrapper;
 use ibc_relayer_runtime::tokio::context::TokioRuntimeContext;
 
 #[derive(Clone)]
