+++ crates/relayer-framework/src/base/one_for_all/impls/relay/types.rs
@@ -2,9 +2,9 @@ use crate::base::chain::types::aliases::{ChannelId, Height, PortId, Sequence, Ti
 use crate::base::core::traits::error::HasErrorType;
 use crate::base::one_for_all::traits::chain::OfaChainTypes;
 use crate::base::one_for_all::traits::relay::OfaBaseRelay;
-use crate::base::one_for_all::traits::runtime::OfaRuntimeWrapper;
 use crate::base::one_for_all::types::chain::OfaChainWrapper;
 use crate::base::one_for_all::types::relay::OfaRelayWrapper;
+use crate::base::one_for_all::types::runtime::OfaRuntimeWrapper;
 use crate::base::relay::traits::types::HasRelayTypes;
 use crate::base::runtime::traits::runtime::HasRuntime;
 use crate::std_prelude::*;
