+++ crates/relayer-framework/src/base/one_for_all/impls/runtime/sleep.rs
@@ -1,7 +1,8 @@
 use async_trait::async_trait;
 use core::time::Duration;
 
-use crate::base::one_for_all::traits::runtime::{OfaRuntime, OfaRuntimeWrapper};
+use crate::base::one_for_all::traits::runtime::OfaRuntime;
+use crate::base::one_for_all::types::runtime::OfaRuntimeWrapper;
 use crate::base::runtime::traits::sleep::CanSleep;
 use crate::std_prelude::*;
