+++ crates/relayer-framework/src/base/one_for_all/impls/runtime/spawn.rs
@@ -1,6 +1,7 @@
 use core::future::Future;
 
-use crate::base::one_for_all::traits::runtime::{OfaRuntime, OfaRuntimeWrapper};
+use crate::base::one_for_all::traits::runtime::OfaRuntime;
+use crate::base::one_for_all::types::runtime::OfaRuntimeWrapper;
 use crate::base::runtime::traits::spawn::{HasSpawner, Spawner};
 use crate::std_prelude::*;
