+++ crates/relayer-framework/src/base/one_for_all/impls/runtime/time.rs
@@ -1,6 +1,7 @@
 use core::time::Duration;
 
-use crate::base::one_for_all::traits::runtime::{OfaRuntime, OfaRuntimeWrapper};
+use crate::base::one_for_all::traits::runtime::OfaRuntime;
+use crate::base::one_for_all::types::runtime::OfaRuntimeWrapper;
 use crate::base::runtime::traits::time::HasTime;
 
 impl<Runtime: OfaRuntime> HasTime for OfaRuntimeWrapper<Runtime> {
