+++ crates/relayer-framework/src/base/one_for_all/traits/relay.rs
@@ -9,9 +9,9 @@ use crate::base::core::traits::sync::Async;
 use crate::base::one_for_all::traits::chain::OfaIbcChainPreset;
 use crate::base::one_for_all::traits::chain::{OfaChainTypes, OfaIbcChain};
 use crate::base::one_for_all::traits::runtime::OfaRuntime;
-use crate::base::one_for_all::traits::runtime::OfaRuntimeWrapper;
 use crate::base::one_for_all::types::chain::OfaChainWrapper;
 use crate::base::one_for_all::types::relay::OfaRelayWrapper;
+use crate::base::one_for_all::types::runtime::OfaRuntimeWrapper;
 use crate::base::relay::traits::ibc_message_sender::IbcMessageSender;
 use crate::base::relay::traits::packet_relayer::PacketRelayer;
 use crate::base::relay::traits::target::{DestinationTarget, SourceTarget};
