+++ crates/relayer-framework/src/base/one_for_all/types/runtime.rs
@@ -1,0 +1,18 @@
+#[derive(Clone)]
+pub struct OfaRuntimeWrapper<Runtime> {
+    pub runtime: Runtime,
+}
+
+impl<Runtime> OfaRuntimeWrapper<Runtime> {
+    pub fn new(runtime: Runtime) -> Self {
+        Self { runtime }
+    }
+}
+
+pub enum LogLevel {
+    Error,
+    Warn,
+    Info,
+    Debug,
+    Trace,
+}
